feat: address concurrency warnings arising from `Theme`
`Theme` was marked as `Sendable` in #351. There's nothing wrong
with this per se, but sendability is not inferred for public
types, a structure is only truely `Sendable` if all of its
members are `Sendable`, and most of the members of `Theme` are
not `Sendable`.

This patch address the sendability warnings for Theme's members
that the compiler emits (regardless of `StrictConcurrency`).